Separate event shadow and peak shadow
